app/src/test/java/com/example/peter/thekitchenmenu/data/repository/source/local/ingredient/dataadapter/IngredientLocalSaveAdapterTest.java
package com.example.peter.thekitchenmenu.data.repository.source.local.ingredient.dataadapter;

import com.example.peter.thekitchenmenu.data.repository.ingredient.TestDataIngredient;
import com.example.peter.thekitchenmenu.data.repository.source.local.ingredient.datasource.IngredientLocalDataSource;
import com.example.peter.thekitchenmenu.data.repository.source.local.ingredient.datasource.TestDataIngredientEntity;
import com.example.peter.thekitchenmenu.domain.usecase.ingredient.IngredientPersistenceModel;

import static org.junit.Assert.*;

import org.junit.Before;
import org.junit.Test;
import org.mockito.ArgumentCaptor;
import org.mockito.Captor;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import static org.junit.Assert.*;
import static org.mockito.Matchers.eq;
import static org.mockito.Mockito.verify;

public class IngredientLocalSaveAdapterTest {

    // region constants ----------------------------------------------------------------------------
    // endregion constants -------------------------------------------------------------------------

    // region helper fields ------------------------------------------------------------------------
    @Mock
    IngredientLocalDataSource repoMock;
    // endregion helper fields ---------------------------------------------------------------------

    private IngredientLocalSaveAdapter SUT;

    @Before
    public void setup() {
        MockitoAnnotations.initMocks(this);
        SUT = givenSystemUnderTest();

    }

    private IngredientLocalSaveAdapter givenSystemUnderTest() {
        return new IngredientLocalSaveAdapter(
                repoMock
        );
    }

    @Test
    public void save() {
        // Arrange
        IngredientPersistenceModel modelUnderTest = TestDataIngredient.
                getValidNewNameValidDescriptionValid();
        // Act
        SUT.save(modelUnderTest);
        // Assert
        verify(repoMock).save(eq(TestDataIngredientEntity.getValidNewNameValidDescriptionValid()));
    }

    // region helper methods -----------------------------------------------------------------------
    // endregion helper methods --------------------------------------------------------------------

    // region helper classes -----------------------------------------------------------------------
    // endregion helper classes --------------------------------------------------------------------
}